# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Projects Service API
  description: |
    Multi-tenant project and task management API for BengoBox platform.

    ## Authentication
    All endpoints (except health checks) require a valid JWT token in the Authorization header.

    ## Multi-tenancy
    All resource endpoints are scoped to a tenant via the `{tenantID}` path parameter.
  version: 1.0.0
  contact:
    name: BengoBox Platform Team
    email: platform@bengobox.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
  - url: https://api.bengobox.com/api/v1
    description: Production
...
paths:
...
  /{tenantID}/tenders:
    parameters:
      - $ref: '#/components/parameters/TenantID'
    get:
      summary: List tenders
      description: Returns a paginated list of tenders for the tenant
      tags: [Tenders]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/TenderStatus'
        - name: priority
          in: query
          schema:
            $ref: '#/components/schemas/Priority'
        - name: source
          in: query
          schema:
            $ref: '#/components/schemas/TenderSource'
        - name: created_by
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of tenders
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListTendersResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create tender
      description: Creates a new tender opportunity with auto-generated tender number
      tags: [Tenders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenderRequest'
      responses:
        '201':
          description: Tender created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenderResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /{tenantID}/tenders/{tenderID}:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - $ref: '#/components/parameters/TenderID'
    get:
      summary: Get tender
      description: Returns a single tender by ID
      tags: [Tenders]
      responses:
        '200':
          description: Tender details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenderResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update tender
      description: Updates an existing tender
      tags: [Tenders]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTenderRequest'
      responses:
        '200':
          description: Tender updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenderResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      summary: Delete tender
      description: Deletes a tender and all associated data
      tags: [Tenders]
      responses:
        '204':
          description: Tender deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
  /{tenantID}/tenders/{tenderID}/meetings/{meetingID}/end:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - $ref: '#/components/parameters/TenderID'
      - $ref: '#/components/parameters/MeetingID'
    post:
      summary: End meeting
      description: Marks a meeting as completed with minutes
      tags: [Tender Meetings]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EndMeetingRequest'
      responses:
        '200':
          description: Meeting ended
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeetingResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /{tenantID}/tenders/{tenderID}/submissions:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - $ref: '#/components/parameters/TenderID'
    get:
      summary: List submissions
      description: Returns submissions for a tender
      tags: [Tender Submissions]
      responses:
        '200':
          description: List of submissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListSubmissionsResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    post:
      summary: Create submission
      description: Creates a new submission record
      tags: [Tender Submissions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubmissionRequest'
      responses:
        '201':
          description: Submission created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
  /{tenantID}/tenders/{tenderID}/submissions/{submissionID}:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - $ref: '#/components/parameters/TenderID'
      - $ref: '#/components/parameters/SubmissionID'
    get:
      summary: Get submission
      description: Returns a submission by ID
      tags: [Tender Submissions]
      responses:
        '200':
          description: Submission details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: Update submission
      description: Updates submission details
      tags: [Tender Submissions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubmissionRequest'
      responses:
        '200':
          description: Submission updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Delete submission
      description: Deletes a submission record
      tags: [Tender Submissions]
      responses:
        '204':
          description: Submission deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /{tenantID}/tenders/{tenderID}/submissions/{submissionID}/submit:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - $ref: '#/components/parameters/TenderID'
      - $ref: '#/components/parameters/SubmissionID'
    post:
      summary: Submit tender
      description: Marks the submission as submitted
      tags: [Tender Submissions]
      responses:
        '200':
          description: Tender submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /{tenantID}/tenders/{tenderID}/submissions/{submissionID}/email-tracking:
    parameters:
      - $ref: '#/components/parameters/TenantID'
      - $ref: '#/components/parameters/TenderID'
      - $ref: '#/components/parameters/SubmissionID'
    post:
      summary: Record email tracking
      description: Records email tracking information
      tags: [Tender Submissions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecordEmailTrackingRequest'
      responses:
        '200':
          description: Email tracking recorded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubmissionResponse'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from auth-service
  parameters:
    TenantID:
      name: tenantID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant identifier
...
    TenderID:
      name: tenderID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Tender identifier
...
    MeetingID:
      name: meetingID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Meeting identifier
...
    SubmissionID:
      name: submissionID
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Submission identifier
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip
  schemas:
...
    Priority:
      type: string
      enum: [low, medium, high, critical]
      description: Priority level
    TenderStatus:
      type: string
      enum: [identified, evaluating, preparing, submitted, under_review, shortlisted, awarded, lost, withdrawn]
      description: Tender lifecycle status
    TenderSource:
      type: string
      enum: [manual, government_portal, referral, website]
      description: Source where tender was discovered
...
    CreateTenderRequest:
      type: object
      required: [title, client_name, deadline]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          example: Government IT Infrastructure Upgrade
        description:
          type: string
          maxLength: 2000
        client_name:
          type: string
          minLength: 1
          maxLength: 255
          example: Ministry of Technology
        client_contact:
          type: string
          maxLength: 255
        client_email:
          type: string
          format: email
        source:
          $ref: '#/components/schemas/TenderSource'
        source_url:
          type: string
          format: uri
        priority:
          $ref: '#/components/schemas/Priority'
        estimated_value:
          type: number
          format: double
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
          example: USD
        publication_date:
          type: string
          format: date-time
        deadline:
          type: string
          format: date-time
        clarification_deadline:
          type: string
          format: date-time
        submission_method:
          $ref: '#/components/schemas/SubmissionMethod'
        submission_address:
          type: string
        categories:
          type: array
          items:
            type: string
        requirements_summary:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
    UpdateTenderRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 2000
        client_name:
          type: string
          minLength: 1
          maxLength: 255
        client_contact:
          type: string
        client_email:
          type: string
          format: email
        source:
          $ref: '#/components/schemas/TenderSource'
        source_url:
          type: string
          format: uri
        status:
          $ref: '#/components/schemas/TenderStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        estimated_value:
          type: number
          format: double
          minimum: 0
        currency:
          type: string
          minLength: 3
          maxLength: 3
        publication_date:
          type: string
          format: date-time
        deadline:
          type: string
          format: date-time
        clarification_deadline:
          type: string
          format: date-time
        submission_method:
          $ref: '#/components/schemas/SubmissionMethod'
        submission_address:
          type: string
        categories:
          type: array
          items:
            type: string
        requirements_summary:
          type: object
          additionalProperties: true
        metadata:
          type: object
          additionalProperties: true
...
    TenderResponse:
      type: object
      required: [id, tenant_id, tender_number, title, client_name, source, status, priority, currency, deadline, created_by, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        tender_number:
          type: string
          example: TND-202501-abc12345
        title:
          type: string
        description:
          type: string
        client_name:
          type: string
        client_contact:
          type: string
        client_email:
          type: string
        source:
          $ref: '#/components/schemas/TenderSource'
        source_url:
          type: string
        status:
          $ref: '#/components/schemas/TenderStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        estimated_value:
          type: number
          format: double
        currency:
          type: string
        publication_date:
          type: string
          format: date-time
        deadline:
          type: string
          format: date-time
        clarification_deadline:
          type: string
          format: date-time
        submission_method:
          $ref: '#/components/schemas/SubmissionMethod'
        submission_address:
          type: string
        categories:
          type: array
          items:
            type: string
        requirements_summary:
          type: object
          additionalProperties: true
        decision:
          type: string
          enum: [go, no_go]
        decision_rationale:
          type: string
        decision_date:
          type: string
          format: date-time
        decided_by:
          type: string
          format: uuid
        project_id:
          type: string
          format: uuid
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
    ListTendersResponse:
      type: object
      required: [tenders, total, limit, offset]
      properties:
        tenders:
          type: array
          items:
            $ref: '#/components/schemas/TenderResponse'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    EndMeetingRequest:
      type: object
      properties:
        minutes:
          type: string
          maxLength: 10000
        decisions:
          type: array
          items:
            type: object
            additionalProperties: true
        action_items:
          type: array
          items:
            type: object
            additionalProperties: true
    MeetingResponse:
      type: object
      required: [id, tenant_id, tender_id, title, meeting_type, scheduled_at, duration_minutes, status, organized_by, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        tender_id:
          type: string
          format: uuid
        committee_id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        meeting_type:
          $ref: '#/components/schemas/MeetingType'
        scheduled_at:
          type: string
          format: date-time
        duration_minutes:
          type: integer
        location:
          type: string
        platform:
          $ref: '#/components/schemas/MeetingPlatform'
        meeting_url:
          type: string
        meeting_id:
          type: string
        status:
          $ref: '#/components/schemas/MeetingStatus'
        attendees:
          type: array
          items:
            type: string
            format: uuid
        agenda:
          type: string
        minutes:
          type: string
        decisions:
          type: array
          items:
            type: object
            additionalProperties: true
        action_items:
          type: array
          items:
            type: object
            additionalProperties: true
        recording_url:
          type: string
        organized_by:
          type: string
          format: uuid
        started_at:
          type: string
          format: date-time
        ended_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
...
    RecordEmailTrackingRequest:
      type: object
      required: [message_id]
      properties:
        message_id:
          type: string
          maxLength: 255
        opened:
          type: boolean
...
    ListSubmissionsResponse:
      type: object
      required: [submissions, total]
      properties:
        submissions:
          type: array
          items:
            $ref: '#/components/schemas/SubmissionResponse'
        total:
          type: integer
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
security:
  - BearerAuth: []